Add a Timeline with play and pause
This commit adds a timeline to control the visualization. It provides
handles for play, pause, 1 frame back, 1 frame forward, beginning and
end.

Additionally this commit adds angularjs to visualize the data retrieved
from the flee simulation. Currently we print all information available
for camps. Thanks to angular this information is updated dynamically
when modifying the timeline.