Add log scaling
This commit scales the data provided using the log function. The data is
scaled before applied to the heatmap visualization and circle
visualization. Additionally the circles now change color based on how
many actors are currently in the given location.

This should help in the future when dealing with general simulations.