Extend support to general simulations
This commit allows the application to visualize any provided simulation
as long as the json format is correct.

The simulation data has to be placed in the directory defined in
config.js under the key dataDirectory. All json files in this directory
will be available for visualization on the web page.

`config.js` was added such that various settings can be changed quickly
and efficiently.