Rework routes and display of circles
This commit updates the visualization of routes and circles.
Routes no longer increase in weight but instead change color to display how many actors are on specific routes.

Circles no longer change color based on hardcoded thresholds. Instead specific thresholds for the current visualization are calculated, based on the maximum amount of actors in a single location.