Add support for different scaling functions
This commit allows the user to select a scaling method from a list of
given scaling methods. Currently we offer logarithmic scaling and a
logistic growth function.

Additionally this commit adds numerous settings associated with the
logistic growth function scaling which can be editet. When changing the
parameters of the function the map is updated live.

Also changes the order and displaying of the map, player bar and
settings window.