New threshold calculation
This commit changes how the thresholds for changing the colors of
circles are calculated. The prior implementation was flawed and let to
wrong results due to weird interactions with scaling mechanisms. The new implementation seperates the data set into thirds and uses this information to find thresholds for which it holds that 30 percent of the data lies below or above them.