Add flexible timestep display
This commit adds the format how a timestep should be displayed in the
visualization to each individual timestep. This means that when the data
is generated it has to specify the date format which is going to be
used.

The benefit is that now you can easily add visualizations that f.e. have
1 minute as each timestep where as before you could only visualize
simulations that used days as timesteps.